112 move management client to separate folder
