+++ src/Kentico.Kontent.ModelGenerator.Core/CodeGenerator.cs
@@ -6,6 +6,7 @@ using Kentico.Kontent.Delivery.Abstractions;
 using Kentico.Kontent.Management.Models.Types;
 using Kentico.Kontent.Management.Models.TypeSnippets;
 using Kentico.Kontent.ModelGenerator.Core.Configuration;
+using Kentico.Kontent.ModelGenerator.Core.ManagementClient;
 using Microsoft.Extensions.Options;
 
 namespace Kentico.Kontent.ModelGenerator.Core
